Исправления после код-ревью

+++ src/main/java/ru/skypro/ads/dto/Comment.java
@@ -2,14 +2,12 @@ package ru.skypro.ads.dto;
 
 import lombok.Data;
 
-import java.util.List;
-
 @Data
 public class Comment {
     private Integer author;
     private String authorImage;
     private String authorFirstName;
-    private Integer createdAt;
+    private Long createdAt;
     private Integer pk;
     private String text;
 }
